MM-18593: support extra plugin metadata
Support parsing `IconPath`, `HomepageURL` and `SupportURL` out of a
plugin's manifest. Embed a timestamp as to when the plugin was last
updated. Update to GitHub 0.12.0 with this metadata.

Note that we maintain the old icons for the existing plugins with icons
for when we expand the database to include all versions available.