MM-18593: support extra plugin metadata (#18)
* temporary commit to point at mm-19477-model-extra-manifest-data

* point GitHub icon at local asset

* MM-18593: support extra plugin metadata

Support parsing `IconPath`, `HomepageURL` and `SupportURL` out of a
plugin's manifest. Embed a timestamp as to when the plugin was last
updated. Update to GitHub 0.12.0 with this metadata.

Note that we maintain the old icons for the existing plugins with icons
for when we expand the database to include all versions available.

* go.mod: mattermost-server@master

* go mod tidy

* remove empty line

* s/getTarFile/getFromTarFile

* make time.UTC explicit

* clarify debug logs

* extra checking

* add support for ReleaseNotesURL

Fixes: https://mattermost.atlassian.net/browse/MM-19602

* split path.Match into two lines